+++ src/app/types/finalityProviders.ts
@@ -17,15 +17,19 @@ export interface Description {
   details: string;
 }
 
-export enum FinalityProviderStatus {
-  FINALITY_PROVIDER_STATUS_ACTIVE = "Active",
-  FINALITY_PROVIDER_STATUS_INACTIVE = "Inactive",
-  FINALITY_PROVIDER_STATUS_JAILED = "Jailed",
-  FINALITY_PROVIDER_STATUS_SLASHED = "Slashed",
+export enum FinalityProviderState {
+  ACTIVE = "FINALITY_PROVIDER_STATUS_ACTIVE",
+  INACTIVE = "FINALITY_PROVIDER_STATUS_INACTIVE",
+  JAILED = "FINALITY_PROVIDER_STATUS_JAILED",
+  SLASHED = "FINALITY_PROVIDER_STATUS_SLASHED",
 }
 
-export type FinalityProviderState =
-  | "FINALITY_PROVIDER_STATUS_ACTIVE"
-  | "FINALITY_PROVIDER_STATUS_INACTIVE"
-  | "FINALITY_PROVIDER_STATUS_JAILED"
-  | "FINALITY_PROVIDER_STATUS_SLASHED";
+export const FinalityProviderStateLabels: Record<
+  FinalityProviderState,
+  string
+> = {
+  [FinalityProviderState.ACTIVE]: "Active",
+  [FinalityProviderState.INACTIVE]: "Inactive",
+  [FinalityProviderState.JAILED]: "Jailed",
+  [FinalityProviderState.SLASHED]: "Slashed",
+};
